Move unit selector to shared folder
